move utils test
